requirements.txt
aiogram~=2.25.1
python-dotenv~=1.0.0
art~=5.8
requests~=2.28.2